aloha-python/client/pyproject.toml
[project]
name = "aloha-python-client"
version = "1.0.0"
description = "Python client implementation with multi-transport support"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.10",
    "grpcio>=1.60.0",
    "httpx>=0.28.1",
    "websockets>=13.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "pytest-asyncio>=0.24.0", "ruff>=0.8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
